Splat Wires Wherever
Ok, so, now we're putting wires on walls, ceilings, and floors.
It's incredible.

They don't _quite_ line up, but that's really an issue of tilesets.
The floors stick up about 1 unit, but the walls don't, the way I've laid things
out in my test center.

Oh well, it's not too bad.
The gist is there.

And along with that, there's an understanding of the true positional "identity"
of the thing I'm pointing at, which should make it easier later to do more
semantic things with the wires. The combination of cube_coord and normal should
work well.

I've had to adjust the wires scene for this new model though.
Now, rather than floating above the floor, to account for that, it's actually
under it. This is because this new identity has the center of the "voxel" as
the origin, and it can then use that to rotate things around it to hit all 6
sides of this space.