Propose Wires
Now the wires are batched. The command is a toggle, which means I point at
things, and when I'm done, they're built.

That's currently worse, but it'll get better later.
What this really allows, though, is intention.
When I go from one place to another, it is no longer "this square has wires"
but instead "this square is connected to this other square"

That will be important later.