WIP: Connecting Wires
So I've added a script for the wires, and made a faded out version of the cross.
Then I made it so it knows where it's going, so things can connect between neighbours.

But... I think I want to move where some things live, and I want a record of
this while it's semi-working, even if it only really works on the floor...