Put Wires on Surfaces
Ok, so, minor hack that makes things easier for me for now.
Now that wires can go everywhere, but only the floor has the offset for now,
I've moved the wires to 5 away from the origin, rather than 4, and made the
code for setting the position of floor tiles aware of the offset.

This makes more sense now that it used to, because I now already had a place
for that logic to live.