Commit Wires
Ok, so bringing it all together, now when I finish drawing wires, they go
solid.
These don't form real connections yet, in that there is no object representing
a "circuit", but they do appear to connect and you can tell the ones you're
proposing from the ones you laid down in the past.

There are still gremlins, for example if you point at a place that doesn't
follow from the previous place, it just kind of allows it.
It should likely not move the head if there isn't a connection.

There's also no way to un-propose a thing.
That will need to be solved eventually, but it won't stand in the way of
learning some things, so I think I'm fine to ignore that for now.