DeadZone
As I was playing around with things I noticed some glitches where traces would
extend out into space sometimes, rather than wrap the proper way around a
corner.

The issue was that sometimes the position of the raytrace would bump down into
another cube, while the normal hadn't changed yet. Or the normal would hit an
edge and do a weird thing.
In both cases, I decided to just ignore the edges of the walls.
No one is really going to be pushing the edges so hard they're mad that it
didn't connect in the first 0.5 units.