Added Circuits
A couple changes here:

First, we put even more logic in the Wires class. Rather than figuring out the
normals and stuff for the proposal on the outside, we instead ask the wire to
please connect itself to the other one.
This is nicer, but also let us add new logic here so Wires can keep track of
their connections.

Then, when I'm committing circuits, I check my neighbours for their circuit,
and adopt what they have.
And if I have multiple neighbours with different circuits, I tell the manager
to merge them.

That leads into my second change. My level is now acting as a circuit manager.
It has a map of the current circuits, it makes new ones, and it merges existing
ones.
Later the circuit will have more state in it, but for now it's just a list of
who's in it.

The third thing is that it's hard to test all of this. So I added a new
collection of meshes to my wires that I can use to tell them apart.
It's not great, because with only 4 shapes it's pretty easy to run into
collisions, but it allowed me to visually test some things and make sure they
behave the way I expect.