Implement Removal Without Splits
Ok, so, this took a bit, but I'm feeling pretty good about it.
Basically, I now have one method that sets the properties of each "leg"

Then I have a state per leg, of nothing, proposed, committed.

Then I set that state when a connection is proposed, but I also look at my
current proposals and the new "proposed deletions" to figure out what the
current state is so I can implement a kind of toggle.

Basically, if I'm being asked to connect two things that are already connected,
then propose we disconnect them.
But if they were already proposed to be disconnected, then add them back together.

Anyway. It seems to work!
But, there are two things I haven't done yet.

First, I don't split circuits. So the disconnected things know they're not
really connected, but they still think they're part of the same circuit.
That's because it seemed like it was going to be some work, and I just wanted
to get this in before I break something.

The second thing is that I'm not removing totally disconnected things.
So I can remove all of a node's connections, and it just hangs out.
I could remove that now, but it felt like the kind of thing I should do after
splitting. Maybe I'm wrong, though.