Added Lamp
This lamp lights up when there's a wire under it, and goes dark when it's
removed.

There's likely more work to be done before this is a re-usable pattern, but for
now it seems to work.

Also, right now it lights up as soon as the proposed wires are beneath it.
I spent some time trying to fix that, but stopped when I realized that
eventually the light will depend on power, not just presence.
So that will take care of that, likely.