Tell Lamp About Circuits
This is a baby step towards powered circuits.
The lamps now know the circuit exists, the wire knows the lamp is attached to
it, and it communicates changes downwards.

Also, my lamp now only runs when there's a circuit. That solves the problem I
had before, since because fully proposed circuits don't have that yet, but
existing ones have a circuit on-hand right away.