Two Generators
I just copied the generator and pasted it a little further along.

This allowed me to verify my expectations, though, that I can have a circuit
that can't be handled by one generator, but can be handled by two.

Then, I can break the circuit and one of the runs od lights will be disabled,
but the other will still shine. Thrilling stuff!