Switches!
I changed my plans a bit here.
Originally I was going to have a switch be on the wire at the connection point
between two wires.
Then it was going to hijack the connection, removing each side from the other,
and inserting itself instead.

But that was getting tricky before I'd even started. How do switches handle 90
degree turns? Maybe I can hijack connections that already exist, but what if it
doesn't yet? Now I have to be watching for other connections going on that I
should steal?
None of it was insurmountable, but it was all a bit annoying.

So instead I've done this. The switch is a style of wire, where the legs are
all disconnected from each other.

This is much simpler to reason about, and more flexible! Now switches can
fan-out, essentially, where only one leg is powered but when the switch is
closed all other legs become powered.
It was also easier to build...

So maybe I'll throw it out someday and do what I was originally planning on,
but for now this will get me on to my next experiment.

In building this, I had to make a `splitCircuit` function, since it's now not
just finishing proposals that split things, opening circuits also does.

I also had to change how circuits were reaped; the naive choice I'd made before
would reap somemthing on the other side of a switch as soon as the switch was
opened, which likely isn't what's desired.