Fixups Connections to Open Switch
Previously wires would still steal or merge the circuit definition of their
neighbours when they were being built, even if the switch was open, and opening
the switch afterwards would have given a different result.

There may be a cleaner way to handle this, but for now this will do.