Change How Things Connect to Wires
Previously we added lights to our scene.
They had an area that was set to detect collisions specifically from a matching
area on wires.  This way, they could tell when a new wire was placed beneath
them and connect to it.

But that lead to a lot of logic about setting positions and rotating and stuff
being duplicated, and I'm planning on making more things that depend on wires
next.

When thinking about it, I realized what I really want is for the things to be
children of the wires. Then the positioning and all of that is handled by the
wire, and I just go along for the ride. The other issue I was facing is having
multiple lights in the same place. For wires I have a big set that holds them
all, but I'd need to make a similar set to hold the other crap, so I could find
existing items and replace them, or whatever. This also solves that, because
now each wire just has one dependent that it's fully in charge of.

So, that made me pretty happy. It cleaned a bunch of things up, and will make
it much easier to add new things in the future. They basically just have to
implement one method now.

But... I also changed how Generators work to be the same, since they were
detecting wires in the same way lamps did.
But unlike lamps, I wanted to place generators from the editor.

So now I have this bootstrap step where I detect that I'm actually a child of
the root rather than a wire, then I have the level build a wire where I am, and
have that instantiate a new generator it can hold onto, while I destroy myself.
Works not too bad, really.