Fixup Issues with Last Approach
These were known issues, but I wanted to make sure I got the last PR in first.
The two issues was that I could now turns a generator into a lightbulb, since
they were now taking up the same slot on a wire.

The second was that when I disconnected a bulb from any circuit, it deleted the
bulb, because it deleted the whole wire. The same thing was true of generators.
So now I don't do delete things that have things on them.