Lower Power Lights
I wasn't happy with how brown-outs felt. It was the same as the switch just
being flipped, and I didn't feel like that was super exciting actually.
And thinking about things I plan on building later, it became even more tricky
to figure out how a totally black-out wouldn't lead to a constant oscillation
between full power and off.

So, when I was imagining what kind of experience I'd want to play, I decided
the answer was the low-power mode. The lights still light, the shields still
shield, the guns still gun, just less and worse.

That way when you're in a hallway and the circuit is overloaded, it's not like
someone just unplugged the whole trunk-line, instead it goes into "oh shit"
mode where you're not literally defense-less, just very very vulnderable. But
that may give you a bit of time to trim power or something.

I played for a while with trying to change the material properties of the bulb,
but eventually just decided to copy what I'd done before and made a copy.
This let me change things in the editor, and it was just faster.

I assume later I'll use one set of lights and bulbs, and use animation to
change from one mode to another, but that's polish. For now it's working, and I
can move on.