First Steps for Turrets
I made a really simple model. It's too big, but I know it's temporary.
So it spins around scanning its area, and if anything targetable hits one of
its feelers, it will point at it.

That's basically it for now, but already there are some issues.
I thought three feelers would be enough, but already I've actually struggled to
be targetted by this thing sometimes. Oh well, I'll fix that later.