Turret Firing
I may have overcomplicated things a bit, but I think I'm still ok.
The gun now has a state machine that goes through various states.

This means the gun now requires power and turns off without it, and after it
locks on to something it begins to fire, which in this case is just a red
cylinder which is made visible.

There's definitely some duplication, and, like I said, perhaps too much state
machine, but I think this is right.

Later I'll add heat and the venting state, but in that case I think this will
be helpful.