Heating and Venting
The gun now accumulates heat while firing, and ditches heat all the time.
If the heat ever goes over the max, it takes a break until the heat can come
all the way back to zero.

That gives the gun a kind of rhythm, and also opens me up for later plans of
offering add-on devices which increase the amount of heat or cooling we see.