Turret Variable Demand
Ok, I think this is the last bit of turret work before there are actually
enemies to defend against.

When the laser is firing it draws more power than when it's scanning or
venting.

This leads to my intended effect, which is that when the circuit is on the edge
of overloaded, it goes over when a badguy approaches.
This is exactly what I wanted. The lights go dim, it's lovely.

When it tuckers itself out, there is a bit of a flicker as I go back into
firing mode, but before it's noticed there's nothing to fire at.
I may be able to fix that by having targetting have less power too, and going
from venting to targetting instead, but that sounds like polish, and for now I
didn't want to split the match case.