Spawn With Direction
I want the shield to be aligned in the direction I'm facing (perpendicular to
it, actually, but either way), so I now need to know what direction the player
was facing when they spawn the shield.

Maybe this will be useful for other things later if I had a spot light or a
turret that didn't just rotate 360 degrees, but for now I'm building it just
for the shield.

Also, since there's currently never any verticality in the plan, we disregard
the vertical component of the camera, and also only have 2 degrees of freedom
when on the floor or ceiling, since the walls should always face up still.

The shields are still doing whatever they already were, but I'm making progress
on getting the generators facing the right direction first.
I've also added a cube so I can tell where it's facing while I'm building this,
and I've disabled the shield spreading so I can focus on this alone.