Fix Normals
I noticed the squares I had added to detect direction were randomly rotated,
not level with the surface they were placed on.

When I added a cylinder to show me where it thought "up" was, it became clear
it had no idea.
So I had to normalize things a bit based on the normal of the actual surface
I'm on.