Spawn Shield Relative to Emitter
Ok, I've added a spawn point to the generator now that controls where the shield is spawned, but very importantly also controls what orientation.

It has a rotation, but nothing inside it. That almost feels dumb until you see
that that's because the shield is drawn in the XZ plane, but the generator
wants it to go XY.

Rather than coding that, though, instead I can just place it in the generator
spatial and use the whole transformation, rather than just the origin.
I can also use the "local axis" button to visualize the transformation.

To be able to see what I'm doing, I temporarily moved one of the grid items out
to make sure "UP" in the shield is "FORWARD" in the generator, like I wanted.