Spread Shield from Emitters
Previously I only spread the shield when it was first created, but now that
more than one emitter can be on the same shield, it is no longer intuitive.

So now I erase all the shields and then seed the shield from the emitters, then
iterate through spreading it from those points.

Then I redo that when I change the emitters, so it's kinda neat like that.