Run Shields on Power
I copied some code from Lamp, so now the shield runs on the grid.

When a generator is created, it does nothing.
But when it gets power, it creates the shield, etc.
When it loses power, it removes itself from the shield.

When a shield loses its last emitter, it cleans itself up.