Layers
I may have screwed this up, but it appears to be working.

I've put the shields on a Shield layer, and the walls and floor on a Walls
layer.
Now, when the shield is looking for neighbours, it only looks at the walls and
the other shields, which is nice. But what's even nicer is that when I cast my
rays for putting objects down and laying circuits, etc, it doesn't see the
shields! That means I can now draw things, and activate switches, through the
shields, which is very helpful.
It also removes a error I was seeing where my ray would hit a shield, think it
was a wall, and make a new emitter in mid-air.

In theory this also paves the way later for turrets and other weapons to fire
through the shield, which could come in handy.
It just means I have to be a bit careful when I add new things to make sure
they impact shields, etc.