Shield Splitting
Let me start by saying after I'd built this, I realized I could have done this
in about 10 minutes if I'd just made a new shield for each emitter, and then
let them spread and merge on their own.
Balls.

This method is likely more computationally efficient, but it was also a lot
more time and complexity, and so also likely more error-prone.
Anyway, in my testing it's looked like it worked.

Basically, shield pieces now know if they represent an emitter, and if so which
one. The emitters also know which shield piece belongs to them. The pieces also
know which other pieces share a border with them.

So the split basically just starts at the first emitter's piece, and it calls
trace on all its neighbours, carrying an array through it all.
If a piece that's being traced represents an emitter, then it adds it to the
list. This is similar but different to the logic I use for tracing circuits.

At the end, I'll have a list of all emitters reachable from this first one. I
consider that a single group, and remove those emitters from my set. If there
are still some emitters left, then these must be disjoint, since they weren't
reached in the first trace. So I do it again to get a second group, etc.

When all emitters are in some group, I make new ones with the new emitters.
It's more convenient than the circuit trace, because I don't actually care
about the pieces. They're disposable. Really I'm just using their connectivity
to figure out how to partition the list of emitters. So that's nice.

Until there's a system for taking damage, I think this is the last thing I
needed to build to make shields "work" the way I wanted.