Enemies: Wandering
Ok, so, I now have my wanderer visiting the various locations it's chosen.
It's dead simple, but it was kinda fun to watch the little guy float around.

And most players likely won't watch him anyway, since later there will be some
sort of detection and a switch into a chase and murder mode.

So I think this satisfies the goals I had of making them not stay in one spot,
and explore a bit, and eventually run up against the players if they do
nothing.