Enemies: Turning and Facing
This isn't a lot of code, but took a surprising amount of "ugh" to get working.
There are a lot of crap implementation details that went in here.

So, I'm not thrilled that every frame I'm building and modifying and computing
a bunch of Vector models, but I don't want to focus on Optimization now.
This is likely not my biggest issue and it probably won't actually impact
playability at this point.

I actually like watching the guy.