Enemies: Simple Vision
Ok, I got a chance to play with `tool` and make a component.
It's... ok!

I set the height and angle and length, and it will make me a "cone" of those
specifications.

It also makes a non-cone shaped box, and then monitors it for things entering
and exiting. And when they enter the non-cone box it checks to see if they're
in the cone.

It's probably awful, but for now I'm pretty happy!