Enemies: Occulded Vision
So now for things inside the vision cone, it makes sure it can trace a line

This is really far from perfect, but since I'm not really building a stealth
game, I think it's good enough...
I have a ray firing straight out from the middle of the detector so because I
was having the main look-point going through the floor when they were far
enough away.

But this isn't great either...
I'm sure it's fine, though.

It allows me to hide around a corner and not get seen.