Enemies: Detection Logic
Ok, I basically just dropped my new vision component into the Wanderer, hooked
up the signals, and it basically just worked. Which is pretty cool.

I made the vision cone transparent so it was easier to see what was going on,
but that wasn't strictly required.

From there it was about what to do with this information.
So I made some new states for chasing a person you can see, going to the last
place you saw a person you were chasing, and taking a look around when you get
there.

The logic is all pretty simple, but to be honest it feels pretty compelling
as-is.
I'm impressed.