Multiple Targets
This is kinda hard to test with the game as-is, but previously I would have
ended up in a situation where if two people walked into view, it would set the
first as the current target and ignore the second, since there already was a
target.
But when the first left, it would have gone into hunt mode even though it
should be able to see another viable target! It would walk right past, since
there are no new events, it was already notified about it entering view, it
just ignored that.

It would keep looking for that first guy until the second one fell out of view
and back in.

So, while it's hard to test, now when one target leaves it _hopefully_ will
just switch to the other. Only if the last target leaves should it go to
hunting them down.

In doing this I had to adjust the order of some things to make sure my state
was right before my signal handlers were called.