Enemies: Animate Face for State
I now show state by the glowing face.
But more specifically, I'm actually testing out using animations.

The key thing here is for the mesh and material both I had to go into the
Resource category and check "Make local to scene", otherwise every instance's
animation changed each other, and it was a disaster.
But with that set, then the animations can go each change a different colour,
and all is good with the world.

This opens me up to do a lot more tricky stuff, I think, so this was really
good practice.