Small Vision Tool Fixup
I have to have this workaround because of the way `tool` scripts handle
reloads, but often when I'd save I'd get warnings telling me I couldn't find
these nodes (I believe it's because I had another tab open, so the object
existed but wasn't in a scene)

Anyway, this is what I actually meant. I didn't know it was going to log an
error the other way.