Fix for Busted Wanderer
Things aren't perfect, and I saw a wanderer get stuck and float up into the air
at some point.
That's obviously bad, but whatever. The real problem was that once they got up
there there were now no more valid places to walk, which gave me an error with
an index out of bounds.

The obviously I don't want it to get into this state, but if it should I would
rather it get stuck than crash the whole game.