Enemies: Fix Wanderer Collision and Vision Layers
I was a bit sloppy earlier, which was fine at the time, but as I come up on a
game, I need to fix it.
The enemies are now on their own layer, so they don't target each other
anymore. Also, they don't go through walls anymore. It _mostly_ didn't matter
before, but it did sometimes come up in edge cases.

For now I've allowed them to walk through each other, but that may change in
the future.