Abstract Heat from Turret
I'm planning on adding heat to the shield emitter soon, so before I do that I
took some of my intentions for Health and moved them into Heat.

I think it worked out nicely. It added a lot more lines than it removed... but
I think that's because of how simple heat is.
It's definitely more work as a Node, but I think I still like it better.