Change Shield Piece from Spatial to Static Body
As I've been going through and adding detection to things I've come to realize
I did things wrong.
Rather than having a spatial that contains a StaticBody, and then having the
collision logic query things about the collided body's parent, I should just be
making these things collidable themselves.

So this simplifies the logic some by making the shield piece itself a body.
This is important, because the next thing I'm going to do is make the shield
piece take damage, and the way something takes damage is by being in the
enemie's kill box and also accepting the method.