Overheat Shield
I've added a heat to the shield emitter. When a shield piece takes damage, it
passes it up to the shield.
When the shield takes damage it divides it among the emitters.

When the emitter detects the shield is taking damage, it absorbs it as heat.

So, when the heat gets to be too much, it disables the shield, same as if it
was unplugged.

It's pretty cool, so far.
There are some issues with the Wanderers being a bit forgetful, so they blow
through the shield and then just chill out, and wander away.

Later, though, the emitter itself will be damaged, and so that will make things
more fair, I think.