Make Shield Emitter Destructable
Ok, so it's mostly pretty easy. Add health, listen for dead signal.
But in doing that, I discovered some issues with my implementation of health...

The max was a constant, so while I could set my own starting value of health,
it didn't matter at all, because it started by setting itself to FULL, which
ignored my value and went right back to 100.

As though that wasn't bad enough, when checking to see if I should change to
FULL, I wasn't even using the constant. I just hard-coded 100!

Anyway, those issues are fixed now, so I can tweak these things per object.

Then there was another issue, which is that if the the emitter and shield were
both in the kill-box then the emitter would get killed out from under the
shield, before the shield had failed.

To fix that, I extended the health module to be able to be invulnerable, and
then I have the shield emitter set itself to invulnerable so long as there's a
shield around.

This is cool, actually, in that it doesn't just set the health to full or
something, so it still heals, it still consumes energy doing so, etc.
If I power-cycle the shield over and over it won't last forever.

It still heals at a normal rate, and takes damage during the moments it's
vulnerable.