Revamp Turret + Replace Rays with Vision
Turrets were a bit busted...
The biggest issue was that I misread the docs, and it lead to a lot of other
weird problems.

The docs for look_at says something like "the local -Z axis", which is pretty
clear in monospace, but in the font the docs were in I thought it was "local-Z"

So it points the Z axis at the thing!
But no, it points the Negative Z axis at the thing.

So that's why I had to use negative source.normal as "up"
And why I had to have the whole thing flipped -180 degrees.
And why my vision was crashing into the floor, rather than sticking way up
above it.

Ok, so I've got the whole thing turned right-way-round now.
My holder doesn't have any rotation.
My barrel just sticks out the other side.
My normal is my normal.
I can use "0" as the angle to reset to when I'm scanning.
And my vision looks like it does in the editor.

While here I also made a tweak to vision.
I was having an issue with the visibility check where it would tilt the whole
thing up to shoot me, but that would move the "level" check to the point where
it couldn't see me.

But for my turret to work, it uses a targeting point.
So now I say if my vision has an aim point, use that, otherwise just use the
origin.
Maybe I broke something, but hopefully I made things better instead...