Turrets Shoot Bad Guys
Two main things.

1) I changed the turret's vision to collide with the enemy layer and not the
player layer

2) I added an aim target so the Wanderer get get shot in it

Also, seeing the laser hit someone else it was clear it was shooting too far.
Turns out I was a bit overzealous with the changes I took out when the whole
thing was upside down and backwards.

The number 2.5 there wasn't just an arbitrary number. It's that number because
the laser lives inside the barrel, and the barrel is 5 units long, and so I
need the laser to move back from the origin of the barrel to be at the place
the ray is firing from.