Wanderers Become Mortal
Ok! This is it!
My turrets can now kill these things!

So I can lure a bad guy, he can optionally be caught in a shield, and meanwhile
a turret can fire on him until he's dead.

It's sorta like a game! Not a great one, but a game nonetheless!