Kinda Like a Game
Ok, I did a few things here.

The biggest is that I added an EnemySpawner.
This produces new bad guys every so often.

Second is I made the turrets a bit smaller.
Now that I want more of them, they were a bit imposing.

I also removed the debug logging of what state the wanderer is in.
When I was watching one walk around it was useful.

Now that there are 20 trying to kill me, it's less useful.

So, this games isn't _fun_ right now. But it does basically make a bunch of bad
guys until you get overrun.

Some observations:

A) Not being able to clear a space really limits me right now. I always planned
   on allowing it, but I think I'm going to have to do that soon.
   Once I've put a turret, I can't really put another one later because the
   first one is still sucking up power.

B) The computer gets even slower with more enemies. Obviously, but it sucks a
   lot

C) Turning shields on and off is tedious. Especially when it gets laggy, you
   flash past the switch, and now your base is full.

D) I noticed some issues with targetting once I started having a turrets on the
   wall. Unsure why, but it hampered my ability to protect myself.

E) To be a real threat, the wanderers need to have a taller killbox. The lights
   on the roof and even shield generators on the roof were all effectively
   invincible.

F) Not a new observation, but going along with the previous, Turrets never had
   health added. This became very clear once things started swarming through my
   shields

G) One tile wide hallways are too narrow for real use.
   Maybe at some point this will be a neat way to provide challenge to people so
   you can see a 1-wide hall and be like "Aw, crap", but that's because it kinda
   sucks.
   I think the main hallways should all be 2.