Allow Clearing of Wires
Ok, I should have built this a long time ago.
This lets me get rid of something I've built.

It's useful for testing stuff, it's useful for regulating my power consumption,
and moving a front-line forwards. And it's useful for if I put something down
which blocks my path, but it turns out I want to walk through there please...