Track Tile State
This is still pretty early days, but it's a concept.
In particular, actually, I think in the final version wires won't actually be a
way to prevent spawning, since then someone could just paint the floor to keep
baddies out.

But this is an idea, at least, and it's easiest to see.