Added Beacon
I think I now have, like, 4 different algorithms for "spreading", but that's
ok...
So anyway! This one goes outwards and finds all spaces up to some limit, and
then adds itself to the tile-space for all the tiles it finds. The way I've
built it it respects walls (or at least walls as I currently have them, which
is a full empty square rather than a divider between two adjacent squares)

And when we turn it off again, they're allowed to be places again.
Magic!