Added Mastermind Auction
Ok, this is the idea I've had bumping around in my head for a while.
So every turn I get some amount of money, and I can spend it on bad guys.

The bad guys have a price based on the density, so it populates up to a point
early on, but it tapers off as the place gets more full.
This is so things don't feel empty, but aren't mad town either, like they used
to be with the spawners and their fixed rate.

I just played a round until I got all the spawn locations to 0, and it actually
sorta worked as a game!
There's a lot of tuning to go. I ran out of power more than I felt like I
should, and the enemies aren't actually a threat, and the turrets are still
invincible, as are the beacons, as is anything on the ceiling.

But it is neat.
The hope is once I start getting more kinds of enemies, there'll be a bit more
variety than I have right now, but that'll be a later thing.