Control Rods
Ok, this is a hack, but that's what I'm doing so it's fine.

Now rather than controlling the robots position directly, I place a rod
when I click. This rod controls the robots, and they stay in formation
wherever I place this thing.

There's no sense of rotation, and the way it's built sucks, but it's not
too bad, actually!