Per Leg Connectivity for Wires
I'm trying to pull switches out of the wires, and also allow for more
complicated switching configurations.

So as part of that I need to configure connectivity per leg, rather than
just all or nothing like I could with the existing switch.

I don't have arbitrary connectivity, just "are you connected to the
middle or not", so I can't, like, bridge two things without having them
touch or anything. At least not yet.

But at least now I could have the left or the right connect to the
middle or something.