Don't Crash on Outside Corners
Due to the weird way walls meet, there's some weird parts at the edges
of corners where it gets an unexpected normal and thinks it's somewhere
it shouldn't be.

The real answer is to fix the way we handle walls and crap in general,
but for now we just don't want it to crash so just make sure I think
these two wires can be connected, and it's good enough for now.