Blueprint Basics
I can now capture the essence of a collection of tiles and then plop
them down somewhere else later.

There's a long way to go, though.
It's currently basically impossible to aim this to put it somewhere I
actually want, and also I can't rotate it.
Also, the way it "merges" with existing wires actually toggles wires
they have in common, which isn't really what I want.
I want or, not xor!

But it's still something.