Ghost Blueprint
I now have a "placing" mode, and when I'm in placing mode I see a
version of my blueprint following my cursor, and when I click it's laid.

The remaining limitations are that I can't rotate them, and I can't
place them on walls / ceilings.

Oh, and there's only one.