Use Ghost Location to Inform Pasting
For now this is a pretty simple change.
I used to take the location of the blueprint and add it to the location
of the prototype within the blueprint to get the final location.

But now that I've got the ghosts, I'm realizing I'm just doing the 3D
engine's work for it. It already traces relative locations through space
for me, so I should just ask it.

I keep a link between the ghost and the prototype it represents, and
when pasting I just get the location it's in and modify that wire.
This should make it harder for these two to diverge, but more important
I'm looking to do rotating next, and I was about to be doing a lot more
of the computer's work for it.