Blueprint Center of Mass
Rather than finding the bottom corner, which is guaranteed to be
furthest from my cursor to the edge of the piece, I'm now picking the
center of mass.

This means it's easier to place, and once I do rotation it'll be at
least sort of rotating around my cursor rather than rotating off into
nowhere.