Blueprint Rotation
That's actually much much easier than I expected!
I just rotated the blueprint around to start, just to see if that part
alone would work, and then I was shocked when it actually fully worked!

It's because of the way I built this. I don't actually tell the wire
which legs to turn on, I tell it which tile it's connected to. This
means when I've rotated the blueprint, and I translate that through to a
global position, the wires figure out themselves which leg is right!

I was worried I'd have to rotate the legs manually or something, but
it's actually just automatic. So good!

Still can't move it between wall and floor, but that's still pretty
good.