Rotating Blueprint Changes Normals
When I've got a blueprint that transitions from the floor to the wall,
it was not working because I'd rotate the wires fine, but they'd end up
on the wrong surface of the other wall.

Now that's working.
I'm still having trouble positioning these things, because they're oddly
placed relative to the cursor, but a least it's working when I can get
it.