Fix Blueprint Placement and Rotation
Previously I picked the center of mass because it was a way to keep the
most tiles in my view while placing as possible.
That made sense, but it turns out it also put me in some impossible
positions where I couldn't put a blueprint in a location because my
center of mass was floating in the middle of space, and I couldn't put
my cursor on that spot, and so I couldn't actually put the tiles there.

Now I find the middlemost tile, but it's always a real tile.
This means I can always find a spot to point at that will put that tile
down, and therefore any valid location could be picked by picking where
that anchor tile would be in this configuration.

That made an issue I'd known about for a while even more noticeable,
though, which is that rotation was broken.
Now that I had an anchor tile it was even more obvious that it wasn't in
the spot I was looking. That because I position tiles by their bottom
corners, not their centers, so when I rotated the tile around it was
rotating around its corner.

Reasonable at a technical level, but felt broken at a human level.
So to fix that I introduced an inner spatial to the blueprint that's
offset by 5,5 so 0,0 of my blueprint is now in the middle of the anchor
tile.

Now when I rotate it rotates the way I want, so I just have to undo that
offset in the setPosition code to make it so it still shows up aligned
with my grid.
I didn't have to change anything about the paste logic, because they use
global_transform so it just keeps working like normal.