Floor Blueprints On Wall
I can now transfer a floor blueprint to the wall, or a wall blueprint to
the floor, or ceiling, etc.

Rotations are also now in the plane of the surface I'm pasting onto, so
when I'm looking at a wall I rotated around the wall's axis, rather than
rotating around the y axis and leaving bits of wire dangling in space.

This means my anchor always stays on the wall, and everything else
rotates to make that work.

I made a bit of a weird choice, and I'm not sure how to feel about it
yet. I keep the original normal, so I'm always finding the minimal
rotation from the original.

This means when I'm going floor to wall it's totally fine.
But when I'm going floor to ceiling I have to pick if I should rotate
one way or another. Currently I'm doing that based on which wall I'm
coming off of, but it means when I go floor to wall to ceiling to other
wall, that last change will be discontinuous, because I'm now rotating
based on what makes the most sense floor to other wall, not ceiling to
other wall. It's weird, but I can always change it later.

Otherwise I may end up in a situation where I go floor, wall1, ceiling,
wall2, floor and end up in a different orientation even though I never
rotated anything.