Turret Tweaks
The turret was my first object, so it fell behind a smidge.
It's still pretty behind, but slightly less so.
It now has health and a hitbox, so it can be destroyed. That's nice,
because it was pretty unfair before.

Also it would sometimes get stuck in a weird state because the vision
pivotted along with the barrel.
This meant that sometimes I would it would see something, but when it
would point it could now see different things because the angles had
changed, etc.

I don't know if I've fully eliminated that, but it's slightly less
prevalent now since I moved the vision out of the barrel's container.
I also made it slightly shorter and narrower so it's not _as_ crazy to
have two of them on two sides of an inside corner.

They still touch a bit, but that's fine for now.

Finally, the laser was on when an unpowered turret was created.
Did I even test this thing?